gravity: Add engine virtual methods and settings file
Add virtual methods in the engine to support children being
able to define what happens during the Loop call.  In this
case, I've added support for a child to define ProcessEvents,
Update, BeginDrawFrame, Draw, and EndDrawFrame.

Hooked up the ability to read a JSON file for the
demo framework settings that define the important details of
the window, engine and application.

Moved engine files into their own folder

Change-Id: I36bb4b7af8bbb573816f93ec5e6cafde80295f26